Version completa del programa

+++ aritmetica.h
@@ -1,10 +1,10 @@
 #ifndef aritmetica_h
 #define aritmetica_h
 
-float suma(float num1, float num2, float* presultado);
-float resta(float num1, float num2, float* resultadoresta);
-float multiplicacion(float num1, float num2, float* resultadomultiplicacion);
-float division(float num1, float num2, float* resultadodivision);
-float factorial_a(float num1, float* resultadofactorial);
-float factorial_b(float num2, float* resultadofactorial);
+int suma(float num1, float num2, float* presultado);
+int resta(float num1, float num2, float* resultadoresta);
+int multiplicacion(float num1, float num2, float* resultadomultiplicacion);
+int division(float num1, float num2, float* resultadodivision);
+int factorial_a(float num1, float* resultadofactorial);
+int factorial_b(float num2, float* resultadofactorial);
 #endif // aritmetica_h
